+++ black_scholes/objects/rho.c
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-int calculate_rho( obj_loc_t obj_loc, int id )
+int calculate_rho( obj_loc_t obj_loc, long id )
 {
     if ( option_call_or_put == CALL )
     {
